Move ProfileDetailsStyle into student profile folder
