+++ app/assets/javascripts/student_profile/ProfileDetailsStyle.js
