+++ ui/legacyHelpers.js
@@ -2,7 +2,6 @@
 import '../app/assets/javascripts/helpers/react_helpers.jsx';
 // (unordered)
 import '../app/assets/javascripts/helpers/mixpanel_utils.jsx';
-import '../app/assets/javascripts/helpers/profile_details_style.jsx';
 import '../app/assets/javascripts/helpers/prop_types.jsx';
 
 // Export these so legacy components can use this, and
